Move validate_error_event_collector_json into validator directory
