+++ tests/agent_features/test_collector_payloads.py
@@ -17,13 +17,15 @@ import webtest
 from testing_support.fixtures import (
     override_application_settings,
     validate_custom_event_collector_json,
-    validate_error_event_collector_json,
     validate_transaction_event_collector_json,
 )
 from testing_support.sample_applications import (
     simple_app,
     simple_custom_event_app,
     simple_exceptional_app,
+)
+from testing_support.validators.validate_error_event_collector_json import (
+    validate_error_event_collector_json,
 )
 from testing_support.validators.validate_error_trace_collector_json import (
     validate_error_trace_collector_json,
