+++ tests/testing_support/fixtures.py
@@ -385,194 +385,6 @@ def make_synthetics_header(account_id, resource_id, job_id, monitor_id, encoding
     return {"X-NewRelic-Synthetics": value}
 
 
-# def validate_transaction_metrics(
-#     name,
-#     group="Function",
-#     background_task=False,
-#     scoped_metrics=None,
-#     rollup_metrics=None,
-#     custom_metrics=None,
-#     index=-1,
-# ):
-#     scoped_metrics = scoped_metrics or []
-#     rollup_metrics = rollup_metrics or []
-#     custom_metrics = custom_metrics or []
-
-#     if background_task:
-#         unscoped_metrics = [
-#             "OtherTransaction/all",
-#             "OtherTransaction/%s/%s" % (group, name),
-#             "OtherTransactionTotalTime",
-#             "OtherTransactionTotalTime/%s/%s" % (group, name),
-#         ]
-#         transaction_scope_name = "OtherTransaction/%s/%s" % (group, name)
-#     else:
-#         unscoped_metrics = [
-#             "WebTransaction",
-#             "WebTransaction/%s/%s" % (group, name),
-#             "WebTransactionTotalTime",
-#             "WebTransactionTotalTime/%s/%s" % (group, name),
-#             "HttpDispatcher",
-#         ]
-#         transaction_scope_name = "WebTransaction/%s/%s" % (group, name)
-
-#     @function_wrapper
-#     def _validate_wrapper(wrapped, instance, args, kwargs):
-
-#         record_transaction_called = []
-#         recorded_metrics = []
-
-#         @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#         @catch_background_exceptions
-#         def _validate_transaction_metrics(wrapped, instance, args, kwargs):
-#             record_transaction_called.append(True)
-#             try:
-#                 result = wrapped(*args, **kwargs)
-#             except:
-#                 raise
-#             else:
-#                 metrics = instance.stats_table
-#                 # Record a copy of the metric value so that the values aren't
-#                 # merged in the future
-#                 _metrics = {}
-#                 for k, v in metrics.items():
-#                     _metrics[k] = copy.copy(v)
-#                 recorded_metrics.append(_metrics)
-
-#             return result
-
-#         def _validate(metrics, name, scope, count):
-#             key = (name, scope)
-#             metric = metrics.get(key)
-
-#             def _metrics_table():
-#                 out = [""]
-#                 out.append("Expected: {0}: {1}".format(key, count))
-#                 for metric_key, metric_value in metrics.items():
-#                     out.append("{0}: {1}".format(metric_key, metric_value[0]))
-#                 return "\n".join(out)
-
-#             def _metric_details():
-#                 return "metric=%r, count=%r" % (key, metric.call_count)
-
-#             if count is not None:
-#                 assert metric is not None, _metrics_table()
-#                 if count == "present":
-#                     assert metric.call_count > 0, _metric_details()
-#                 else:
-#                     assert metric.call_count == count, _metric_details()
-
-#                 assert metric.total_call_time >= 0, (key, metric)
-#                 assert metric.total_exclusive_call_time >= 0, (key, metric)
-#                 assert metric.min_call_time >= 0, (key, metric)
-#                 assert metric.sum_of_squares >= 0, (key, metric)
-
-#             else:
-#                 assert metric is None, _metrics_table()
-
-#         _new_wrapper = _validate_transaction_metrics(wrapped)
-#         val = _new_wrapper(*args, **kwargs)
-#         assert record_transaction_called
-#         metrics = recorded_metrics[index]
-
-#         record_transaction_called[:] = []
-#         recorded_metrics[:] = []
-
-#         for unscoped_metric in unscoped_metrics:
-#             _validate(metrics, unscoped_metric, "", 1)
-
-#         for scoped_name, scoped_count in scoped_metrics:
-#             _validate(metrics, scoped_name, transaction_scope_name, scoped_count)
-
-#         for rollup_name, rollup_count in rollup_metrics:
-#             _validate(metrics, rollup_name, "", rollup_count)
-
-#         for custom_name, custom_count in custom_metrics:
-#             _validate(metrics, custom_name, "", custom_count)
-
-#         custom_metric_names = {name for name, _ in custom_metrics}
-#         for name, _ in metrics:
-#             if name not in custom_metric_names:
-#                 assert not name.startswith("Supportability/api/"), name
-
-#         return val
-
-#     return _validate_wrapper
-
-
-# def validate_time_metrics_outside_transaction(time_metrics=None, index=-1):
-#     time_metrics = time_metrics or []
-
-#     @function_wrapper
-#     def _validate_wrapper(wrapped, instance, args, kwargs):
-
-#         record_time_metric_called = []
-#         recorded_metrics = []
-
-#         @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_time_metric")
-#         @catch_background_exceptions
-#         def _validate_transaction_metrics(wrapped, instance, args, kwargs):
-#             record_time_metric_called.append(True)
-#             try:
-#                 result = wrapped(*args, **kwargs)
-#             except:
-#                 raise
-#             else:
-#                 metrics = instance.stats_table
-#                 # Record a copy of the metric value so that the values aren't
-#                 # merged in the future
-#                 _metrics = {}
-#                 for k, v in metrics.items():
-#                     _metrics[k] = copy.copy(v)
-#                 recorded_metrics.append(_metrics)
-
-#             return result
-
-#         def _validate(metrics, name, count):
-#             key = (name, "")
-#             metric = metrics.get(key)
-
-#             def _metrics_table():
-#                 out = [""]
-#                 out.append("Expected: {0}: {1}".format(key, count))
-#                 for metric_key, metric_value in metrics.items():
-#                     out.append("{0}: {1}".format(metric_key, metric_value[0]))
-#                 return "\n".join(out)
-
-#             def _metric_details():
-#                 return "metric=%r, count=%r" % (key, metric.call_count)
-
-#             if count is not None:
-#                 assert metric is not None, _metrics_table()
-#                 if count == "present":
-#                     assert metric.call_count > 0, _metric_details()
-#                 else:
-#                     assert metric.call_count == count, _metric_details()
-
-#                 assert metric.total_call_time >= 0, (key, metric)
-#                 assert metric.total_exclusive_call_time >= 0, (key, metric)
-#                 assert metric.min_call_time >= 0, (key, metric)
-#                 assert metric.sum_of_squares >= 0, (key, metric)
-
-#             else:
-#                 assert metric is None, _metrics_table()
-
-#         _new_wrapper = _validate_transaction_metrics(wrapped)
-#         val = _new_wrapper(*args, **kwargs)
-#         assert record_time_metric_called
-#         metrics = recorded_metrics[index]
-
-#         record_time_metric_called[:] = []
-#         recorded_metrics[:] = []
-
-#         for time_metric, count in time_metrics:
-#             _validate(metrics, time_metric, count)
-
-#         return val
-
-#     return _validate_wrapper
-
-
 def capture_transaction_metrics(metrics_list, full_metrics=None):
     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
     @catch_background_exceptions
@@ -592,264 +404,6 @@ def capture_transaction_metrics(metrics_list, full_metrics=None):
         return result
 
     return _capture_transaction_metrics
-
-
-# def validate_internal_metrics(metrics=None):
-#     metrics = metrics or []
-
-#     def no_op(wrapped, instance, args, kwargs):
-#         pass
-
-#     @function_wrapper
-#     def _validate_wrapper(wrapped, instance, args, kwargs):
-#         # Apply no-op wrappers to prevent new internal trace contexts from being started, preventing capture
-#         wrapped = transient_function_wrapper("newrelic.core.internal_metrics", "InternalTraceContext.__enter__")(no_op)(wrapped)
-#         wrapped = transient_function_wrapper("newrelic.core.internal_metrics", "InternalTraceContext.__exit__")(no_op)(wrapped)
-
-#         captured_metrics = CustomMetrics()
-#         with InternalTraceContext(captured_metrics):
-#             result = wrapped(*args, **kwargs)
-#         captured_metrics = dict(captured_metrics.metrics())
-
-#         def _validate(name, count):
-#             metric = captured_metrics.get(name)
-
-#             def _metrics_table():
-#                 return "metric=%r, metrics=%r" % (name, captured_metrics)
-
-#             def _metric_details():
-#                 return "metric=%r, count=%r" % (name, metric.call_count)
-
-#             if count is not None and count > 0:
-#                 assert metric is not None, _metrics_table()
-#                 if count == "present":
-#                     assert metric.call_count > 0, _metric_details()
-#                 else:
-#                     assert metric.call_count == count, _metric_details()
-
-#             else:
-#                 assert metric is None, _metrics_table()
-
-#         for metric, count in metrics:
-#             _validate(metric, count)
-
-#         return result
-
-#     return _validate_wrapper
-
-
-# def validate_transaction_errors(errors=None, required_params=None, forgone_params=None):
-#     errors = errors or []
-#     required_params = required_params or []
-#     forgone_params = forgone_params or []
-#     captured_errors = []
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     @catch_background_exceptions
-#     def _capture_transaction_errors(wrapped, instance, args, kwargs):
-#         def _bind_params(transaction, *args, **kwargs):
-#             return transaction
-
-#         transaction = _bind_params(*args, **kwargs)
-#         captured = transaction.errors
-
-#         captured_errors.append(captured)
-
-#         return wrapped(*args, **kwargs)
-
-#     @function_wrapper
-#     def _validate_transaction_errors(wrapped, instance, args, kwargs):
-#         _new_wrapped = _capture_transaction_errors(wrapped)
-#         output = _new_wrapped(*args, **kwargs)
-
-#         expected = sorted(errors)
-
-#         if captured_errors:
-#             captured = captured_errors[0]
-#         else:
-#             captured = []
-
-#         if errors and isinstance(errors[0], (tuple, list)):
-#             compare_to = sorted([(e.type, e.message) for e in captured])
-#         else:
-#             compare_to = sorted([e.type for e in captured])
-
-#         assert expected == compare_to, "expected=%r, captured=%r, errors=%r" % (expected, compare_to, captured)
-
-#         for e in captured:
-#             assert e.span_id
-#             for name, value in required_params:
-#                 assert name in e.custom_params, "name=%r, params=%r" % (name, e.custom_params)
-#                 assert e.custom_params[name] == value, "name=%r, value=%r, params=%r" % (
-#                     name,
-#                     value,
-#                     e.custom_params,
-#                 )
-
-#             for name, value in forgone_params:
-#                 assert name not in e.custom_params, "name=%r, params=%r" % (name, e.custom_params)
-
-#         return output
-
-#     return _validate_transaction_errors
-
-
-# def validate_application_errors(errors=None, required_params=None, forgone_params=None):
-#     errors = errors or []
-#     required_params = required_params or []
-#     forgone_params = forgone_params or []
-
-#     @function_wrapper
-#     def _validate_application_errors(wrapped, instace, args, kwargs):
-
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             stats = core_application_stats_engine()
-
-#             app_errors = stats.error_data()
-
-#             expected = sorted(errors)
-#             captured = sorted([(e.type, e.message) for e in stats.error_data()])
-
-#             assert expected == captured, "expected=%r, captured=%r, errors=%r" % (expected, captured, app_errors)
-
-#             for e in app_errors:
-#                 for name, value in required_params:
-#                     assert name in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
-#                     assert e.parameters["userAttributes"][name] == value, "name=%r, value=%r, params=%r" % (
-#                         name,
-#                         value,
-#                         e.parameters,
-#                     )
-
-#                 for name, value in forgone_params:
-#                     assert name not in e.parameters["userAttributes"], "name=%r, params=%r" % (name, e.parameters)
-
-#         return result
-
-#     return _validate_application_errors
-
-
-# def validate_custom_parameters(required_params=None, forgone_params=None):
-#     required_params = required_params or []
-#     forgone_params = forgone_params or []
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     @catch_background_exceptions
-#     def _validate_custom_parameters(wrapped, instance, args, kwargs):
-#         def _bind_params(transaction, *args, **kwargs):
-#             return transaction
-
-#         transaction = _bind_params(*args, **kwargs)
-
-#         # these are pre-destination applied attributes, so they may not
-#         # actually end up in a transaction/error trace, we are merely testing
-#         # for presence on the TransactionNode
-
-#         attrs = {}
-#         for attr in transaction.user_attributes:
-#             attrs[attr.name] = attr.value
-
-#         for name, value in required_params:
-#             assert name in attrs, "name=%r, params=%r" % (name, attrs)
-#             assert attrs[name] == value, "name=%r, value=%r, params=%r" % (name, value, attrs)
-
-#         for name, value in forgone_params:
-#             assert name not in attrs, "name=%r, params=%r" % (name, attrs)
-
-#         return wrapped(*args, **kwargs)
-
-#     return _validate_custom_parameters
-
-
-# def validate_synthetics_event(required_attrs=None, forgone_attrs=None, should_exist=True):
-#     required_attrs = required_attrs or []
-#     forgone_attrs = forgone_attrs or []
-#     failed = []
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     def _validate_synthetics_event(wrapped, instance, args, kwargs):
-#         result = wrapped(*args, **kwargs)
-
-#         try:
-#             if not should_exist:
-#                 assert instance.synthetics_events == []
-#             else:
-#                 assert len(instance.synthetics_events) == 1
-#                 event = instance.synthetics_events[0]
-#                 assert event is not None
-#                 assert len(event) == 3
-
-#                 def _flatten(event):
-#                     result = {}
-#                     for elem in event:
-#                         for k, v in elem.items():
-#                             result[k] = v
-#                     return result
-
-#                 flat_event = _flatten(event)
-
-#                 assert "nr.guid" in flat_event, "name=%r, event=%r" % ("nr.guid", flat_event)
-
-#                 for name, value in required_attrs:
-#                     assert name in flat_event, "name=%r, event=%r" % (name, flat_event)
-#                     assert flat_event[name] == value, "name=%r, value=%r, event=%r" % (name, value, flat_event)
-
-#                 for name, value in forgone_attrs:
-#                     assert name not in flat_event, "name=%r, value=%r, event=%r" % (name, value, flat_event)
-#         except Exception as e:
-#             failed.append(e)
-
-#         return result
-
-#     @function_wrapper
-#     def wrapper(wrapped, instance, args, kwargs):
-#         _new_wrapper = _validate_synthetics_event(wrapped)
-#         result = _new_wrapper(*args, **kwargs)
-#         if failed:
-#             e = failed.pop()
-#             raise e
-#         return result
-
-#     return wrapper
-
-
-# def validate_transaction_event_attributes(required_params=None, forgone_params=None, exact_attrs=None, index=-1):
-#     required_params = required_params or {}
-#     forgone_params = forgone_params or {}
-#     exact_attrs = exact_attrs or {}
-
-#     captured_events = []
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     def _capture_transaction_events(wrapped, instance, args, kwargs):
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-#             event_data = instance.transaction_events
-#             captured_events.append(event_data)
-#             return result
-
-#     @function_wrapper
-#     def _validate_transaction_event_attributes(wrapped, instance, args, kwargs):
-#         _new_wrapper = _capture_transaction_events(wrapped)
-#         result = _new_wrapper(*args, **kwargs)
-
-#         assert captured_events, "No events captured"
-#         event_data = captured_events[index]
-#         captured_events[:] = []
-
-#         check_event_attributes(event_data, required_params, forgone_params, exact_attrs)
-
-#         return result
-
-#     return _validate_transaction_event_attributes
 
 
 def check_event_attributes(event_data, required_params=None, forgone_params=None, exact_attrs=None):
@@ -887,390 +441,6 @@ def check_event_attributes(event_data, required_params=None, forgone_params=None
             assert user_attributes[param] == value, ((param, value), user_attributes)
         for param, value in exact_attrs["intrinsic"].items():
             assert intrinsics[param] == value, ((param, value), intrinsics)
-
-
-# def validate_non_transaction_error_event(required_intrinsics=None, num_errors=1, required_user=None, forgone_user=None):
-#     """Validate error event data for a single error occurring outside of a
-#     transaction.
-#     """
-#     required_intrinsics = required_intrinsics or {}
-#     required_user = required_user or {}
-#     forgone_user = forgone_user or []
-
-#     @function_wrapper
-#     def _validate_non_transaction_error_event(wrapped, instace, args, kwargs):
-
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             stats = core_application_stats_engine(None)
-
-#             assert stats.error_events.num_seen == num_errors
-#             for event in stats.error_events:
-
-#                 assert len(event) == 3  # [intrinsic, user, agent attributes]
-
-#                 intrinsics = event[0]
-
-#                 # The following attributes are all required, and also the only
-#                 # intrinsic attributes that can be included in an error event
-#                 # recorded outside of a transaction
-
-#                 assert intrinsics["type"] == "TransactionError"
-#                 assert intrinsics["transactionName"] is None
-#                 assert intrinsics["error.class"] == required_intrinsics["error.class"]
-#                 assert intrinsics["error.message"].startswith(required_intrinsics["error.message"])
-#                 assert intrinsics["error.expected"] == required_intrinsics["error.expected"]
-#                 now = time.time()
-#                 assert isinstance(intrinsics["timestamp"], int)
-#                 assert intrinsics["timestamp"] <= 1000.0 * now
-
-#                 user_params = event[1]
-#                 for name, value in required_user.items():
-#                     assert name in user_params, "name=%r, params=%r" % (name, user_params)
-#                     assert user_params[name] == value, "name=%r, value=%r, params=%r" % (name, value, user_params)
-
-#                 for param in forgone_user:
-#                     assert param not in user_params
-
-#         return result
-
-#     return _validate_non_transaction_error_event
-
-
-# def validate_application_error_trace_count(num_errors):
-#     """Validate error event data for a single error occurring outside of a
-#     transaction.
-#     """
-
-#     @function_wrapper
-#     def _validate_application_error_trace_count(wrapped, instace, args, kwargs):
-
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             stats = core_application_stats_engine(None)
-#             assert len(stats.error_data()) == num_errors
-
-#         return result
-
-#     return _validate_application_error_trace_count
-
-
-# def validate_application_error_event_count(num_errors):
-#     """Validate error event data for a single error occurring outside of a
-#     transaction.
-#     """
-
-#     @function_wrapper
-#     def _validate_application_error_event_count(wrapped, instace, args, kwargs):
-
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             stats = core_application_stats_engine(None)
-#             assert len(list(stats.error_events)) == num_errors
-
-#         return result
-
-#     return _validate_application_error_event_count
-
-
-# def validate_synthetics_transaction_trace(required_params=None, forgone_params=None, should_exist=True):
-#     required_params = required_params or {}
-#     forgone_params = forgone_params or {}
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     def _validate_synthetics_transaction_trace(wrapped, instance, args, kwargs):
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             # Now that transaction has been recorded, generate
-#             # a transaction trace
-
-#             connections = SQLConnections()
-#             trace_data = instance.transaction_trace_data(connections)
-
-#             # Check that synthetics resource id is in TT header
-
-#             header = trace_data[0]
-#             header_key = "synthetics_resource_id"
-
-#             if should_exist:
-#                 assert header_key in required_params
-#                 assert header[9] == required_params[header_key], "name=%r, header=%r" % (header_key, header)
-#             else:
-#                 assert header[9] is None
-
-#             # Check that synthetics ids are in TT custom params
-
-#             pack_data = unpack_field(trace_data[0][4])
-#             tt_intrinsics = pack_data[0][4]["intrinsics"]
-
-#             for name in required_params:
-#                 assert name in tt_intrinsics, "name=%r, intrinsics=%r" % (name, tt_intrinsics)
-#                 assert tt_intrinsics[name] == required_params[name], "name=%r, value=%r, intrinsics=%r" % (
-#                     name,
-#                     required_params[name],
-#                     tt_intrinsics,
-#                 )
-
-#             for name in forgone_params:
-#                 assert name not in tt_intrinsics, "name=%r, intrinsics=%r" % (name, tt_intrinsics)
-
-#         return result
-
-#     return _validate_synthetics_transaction_trace
-
-
-# def validate_tt_collector_json(
-#     required_params=None,
-#     forgone_params=None,
-#     should_exist=True,
-#     datastore_params=None,
-#     datastore_forgone_params=None,
-#     message_broker_params=None,
-#     message_broker_forgone_params=None,
-#     exclude_request_uri=False,
-# ):
-#     """make assertions based off the cross-agent spec on transaction traces"""
-#     required_params = required_params or {}
-#     forgone_params = forgone_params or {}
-#     datastore_params = datastore_params or {}
-#     datastore_forgone_params = datastore_forgone_params or {}
-#     message_broker_params = message_broker_params or {}
-#     message_broker_forgone_params = message_broker_forgone_params or []
-
-#     @function_wrapper
-#     def _validate_wrapper(wrapped, instance, args, kwargs):
-
-#         traces_recorded = []
-
-#         @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#         def _validate_tt_collector_json(wrapped, instance, args, kwargs):
-
-#             result = wrapped(*args, **kwargs)
-
-#             # Now that transaction has been recorded, generate
-#             # a transaction trace
-
-#             connections = SQLConnections()
-#             trace_data = instance.transaction_trace_data(connections)
-#             traces_recorded.append(trace_data)
-
-#             return result
-
-#         def _validate_trace(trace):
-#             assert isinstance(trace[0], float)  # absolute start time (ms)
-#             assert isinstance(trace[1], float)  # duration (ms)
-#             assert trace[0] > 0  # absolute time (ms)
-#             assert isinstance(trace[2], six.string_types)  # transaction name
-#             if trace[2].startswith("WebTransaction"):
-#                 if exclude_request_uri:
-#                     assert trace[3] is None  # request url
-#                 else:
-#                     assert isinstance(trace[3], six.string_types)
-#                     # query parameters should not be captured
-#                     assert "?" not in trace[3]
-
-#             # trace details -- python agent always uses condensed trace array
-
-#             trace_details, string_table = unpack_field(trace[4])
-#             assert len(trace_details) == 5
-#             assert isinstance(trace_details[0], float)  # start time (ms)
-
-#             # the next two items should be empty dicts, old parameters stuff,
-#             # placeholders for now
-
-#             assert isinstance(trace_details[1], dict)
-#             assert len(trace_details[1]) == 0
-#             assert isinstance(trace_details[2], dict)
-#             assert len(trace_details[2]) == 0
-
-#             # root node in slot 3
-
-#             root_node = trace_details[3]
-#             assert isinstance(root_node[0], float)  # entry timestamp
-#             assert isinstance(root_node[1], float)  # exit timestamp
-#             assert root_node[2] == "ROOT"
-#             assert isinstance(root_node[3], dict)
-#             assert len(root_node[3]) == 0  # spec shows empty (for root)
-#             children = root_node[4]
-#             assert isinstance(children, list)
-
-#             # there are two optional items at the end of trace segments,
-#             # class name that segment is in, and method name function is in;
-#             # Python agent does not use these (only Java does)
-
-#             # let's just test the first child
-#             trace_segment = children[0]
-#             assert isinstance(trace_segment[0], float)  # entry timestamp
-#             assert isinstance(trace_segment[1], float)  # exit timestamp
-#             assert isinstance(trace_segment[2], six.string_types)  # scope
-#             assert isinstance(trace_segment[3], dict)  # request params
-#             assert isinstance(trace_segment[4], list)  # children
-
-#             assert trace_segment[0] >= root_node[0]  # trace starts after root
-
-#             def _check_params_and_start_time(node):
-#                 children = node[4]
-#                 for child in children:
-#                     assert child[0] >= node[0]  # child started after parent
-#                     _check_params_and_start_time(child)
-
-#                 params = node[3]
-#                 assert isinstance(params, dict)
-
-#                 # We should always report exclusive_duration_millis on a
-#                 # segment. This allows us to override exclusive time
-#                 # calculations on APM.
-#                 assert "exclusive_duration_millis" in params
-#                 assert isinstance(params["exclusive_duration_millis"], float)
-
-#                 segment_name = _lookup_string_table(node[2], string_table, default=node[2])
-#                 if segment_name.startswith("Datastore"):
-#                     for key in datastore_params:
-#                         assert key in params, key
-#                         assert params[key] == datastore_params[key]
-#                     for key in datastore_forgone_params:
-#                         assert key not in params, key
-
-#                     # if host is reported, it cannot be localhost
-#                     if "host" in params:
-#                         assert params["host"] not in LOCALHOST_EQUIVALENTS
-
-#                 elif segment_name.startswith("MessageBroker"):
-#                     for key in message_broker_params:
-#                         assert key in params, key
-#                         assert params[key] == message_broker_params[key]
-#                     for key in message_broker_forgone_params:
-#                         assert key not in params, key
-
-#             _check_params_and_start_time(trace_segment)
-
-#             attributes = trace_details[4]
-
-#             assert "intrinsics" in attributes
-#             assert "userAttributes" in attributes
-#             assert "agentAttributes" in attributes
-
-#             assert isinstance(trace[5], six.string_types)  # GUID
-#             assert trace[6] is None  # reserved for future use
-#             assert trace[7] is False  # deprecated force persist flag
-
-#             # x-ray session ID
-
-#             assert trace[8] is None
-
-#             # Synthetics ID
-
-#             assert trace[9] is None or isinstance(trace[9], six.string_types)
-
-#             assert isinstance(string_table, list)
-#             for name in string_table:
-#                 assert isinstance(name, six.string_types)  # metric name
-
-#         _new_wrapper = _validate_tt_collector_json(wrapped)
-#         val = _new_wrapper(*args, **kwargs)
-#         trace_data = traces_recorded.pop()
-#         trace = trace_data[0]  # 1st trace
-#         _validate_trace(trace)
-#         return val
-
-#     return _validate_wrapper
-
-
-# def validate_transaction_trace_attributes(
-#     required_params=None, forgone_params=None, should_exist=True, url=None, index=-1
-# ):
-#     required_params = required_params or {}
-#     forgone_params = forgone_params or {}
-
-#     trace_data = []
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     def _validate_transaction_trace_attributes(wrapped, instance, args, kwargs):
-
-#         result = wrapped(*args, **kwargs)
-
-#         # Now that transaction has been recorded, generate
-#         # a transaction trace
-
-#         connections = SQLConnections()
-#         _trace_data = instance.transaction_trace_data(connections)
-#         trace_data.append(_trace_data)
-
-#         return result
-
-#     @function_wrapper
-#     def wrapper(wrapped, instance, args, kwargs):
-#         _new_wrapper = _validate_transaction_trace_attributes(wrapped)
-#         result = _new_wrapper(*args, **kwargs)
-
-#         _trace_data = trace_data[index]
-#         trace_data[:] = []
-
-#         if url is not None:
-#             trace_url = _trace_data[0][3]
-#             assert url == trace_url
-
-#         pack_data = unpack_field(_trace_data[0][4])
-#         assert len(pack_data) == 2
-#         assert len(pack_data[0]) == 5
-#         parameters = pack_data[0][4]
-
-#         assert "intrinsics" in parameters
-#         assert "userAttributes" in parameters
-#         assert "agentAttributes" in parameters
-
-#         check_attributes(parameters, required_params, forgone_params)
-
-#         return result
-
-#     return wrapper
-
-
-# def validate_transaction_error_trace_attributes(required_params=None, forgone_params=None, exact_attrs=None):
-#     """Check the error trace for attributes, expect only one error to be
-#     present in the transaction.
-#     """
-#     required_params = required_params or {}
-#     forgone_params = forgone_params or {}
-#     exact_attrs = exact_attrs or {}
-
-#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-#     def _validate_transaction_error_trace(wrapped, instance, args, kwargs):
-#         try:
-#             result = wrapped(*args, **kwargs)
-#         except:
-#             raise
-#         else:
-
-#             error_data = instance.error_data()
-
-#             # there should be only one error
-#             assert len(error_data) == 1
-#             traced_error = error_data[0]
-
-#             check_error_attributes(
-#                 traced_error.parameters, required_params, forgone_params, exact_attrs, is_transaction=True
-#             )
-
-#         return result
-
-#     return _validate_transaction_error_trace
 
 
 def check_error_attributes(
@@ -1369,55 +539,55 @@ def check_attributes(parameters, required_params=None, forgone_params=None, exac
 #     return _validate_error_trace_collector_json
 
 
-def validate_error_event_collector_json(num_errors=1):
-    """Validate the format, types and number of errors of the data we
-    send to the collector for harvest.
-    """
+# def validate_error_event_collector_json(num_errors=1):
+#     """Validate the format, types and number of errors of the data we
+#     send to the collector for harvest.
+#     """
 
-    @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
-    def _validate_error_event_collector_json(wrapped, instance, args, kwargs):
-        try:
-            result = wrapped(*args, **kwargs)
-        except:
-            raise
-        else:
+#     @transient_function_wrapper("newrelic.core.stats_engine", "StatsEngine.record_transaction")
+#     def _validate_error_event_collector_json(wrapped, instance, args, kwargs):
+#         try:
+#             result = wrapped(*args, **kwargs)
+#         except:
+#             raise
+#         else:
 
-            samples = list(instance.error_events)
-            s_info = instance.error_events.sampling_info
-            agent_run_id = 666
+#             samples = list(instance.error_events)
+#             s_info = instance.error_events.sampling_info
+#             agent_run_id = 666
 
-            # emulate the payload used in data_collector.py
+#             # emulate the payload used in data_collector.py
 
-            payload = (agent_run_id, s_info, samples)
-            collector_json = json_encode(payload)
+#             payload = (agent_run_id, s_info, samples)
+#             collector_json = json_encode(payload)
 
-            decoded_json = json.loads(collector_json)
+#             decoded_json = json.loads(collector_json)
 
-            assert decoded_json[0] == agent_run_id
+#             assert decoded_json[0] == agent_run_id
 
-            sampling_info = decoded_json[1]
+#             sampling_info = decoded_json[1]
 
-            harvest_config = instance.settings.event_harvest_config
-            reservoir_size = harvest_config.harvest_limits.error_event_data
+#             harvest_config = instance.settings.event_harvest_config
+#             reservoir_size = harvest_config.harvest_limits.error_event_data
 
-            assert sampling_info["reservoir_size"] == reservoir_size
-            assert sampling_info["events_seen"] == num_errors
+#             assert sampling_info["reservoir_size"] == reservoir_size
+#             assert sampling_info["events_seen"] == num_errors
 
-            error_events = decoded_json[2]
+#             error_events = decoded_json[2]
 
-            assert len(error_events) == num_errors
-            for event in error_events:
+#             assert len(error_events) == num_errors
+#             for event in error_events:
 
-                # event is an array containing intrinsics, user-attributes,
-                # and agent-attributes
+#                 # event is an array containing intrinsics, user-attributes,
+#                 # and agent-attributes
 
-                assert len(event) == 3
-                for d in event:
-                    assert isinstance(d, dict)
+#                 assert len(event) == 3
+#                 for d in event:
+#                     assert isinstance(d, dict)
 
-        return result
+#         return result
 
-    return _validate_error_event_collector_json
+#     return _validate_error_event_collector_json
 
 
 def validate_transaction_event_collector_json():
